cloud_functions/carga_incremental_api/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
datetime
fsspec
gcsfs
pandas-gbq